Support creating/deleting/upgrading groups and communities (#7880)

+++ backend/canisters/user_index/impl/src/lib.rs
@@ -264,8 +264,8 @@ impl RuntimeState {
             },
             canister_upgrades_completed: canister_upgrades_metrics.completed,
             canister_upgrades_failed: canister_upgrades_metrics.failed,
-            canister_upgrades_pending: canister_upgrades_metrics.pending as u64,
-            canister_upgrades_in_progress: canister_upgrades_metrics.in_progress as u64,
+            canister_upgrades_pending: canister_upgrades_metrics.pending,
+            canister_upgrades_in_progress: canister_upgrades_metrics.in_progress,
             governance_principals: self.data.governance_principals.iter().copied().collect(),
             user_wasm_version: self.data.child_canister_wasms.get(ChildCanisterType::User).wasm.version,
             local_user_index_wasm_version: self
